Unit test checkpoint. Data type tests and armored data tests completed round 1.

+++ src/org/cryptokitty/digest/HashFactory.java
@@ -27,7 +27,7 @@ public class HashFactory {
 	/*
 	 * Return a message digest given an algorithm constant.
 	 */
-	public static Hash getDigest(byte algorithm)
+	public static Hash getDigest(int algorithm)
 			throws UnsupportedAlgorithmException {
 
 		try {
